requirements.txt
torch==2.0.1
torchvision==0.15.2
gcsfs==2023.6.0
PyYAML==6.0
requests==2.28.1
jq
